Add TravisCI
NB the build will almost certainly fail.  This is probably because there
is a problem installing Bio::Sam::DB.

I've discussed this with @andrewjpage and we think that the module isn't
required because there is already a sam parser included from another
module and that this should be refactored out.  In the mean time I'll
raise a ticket and leave the build red.

There is always a chance that there is a bug in
[install_dependencies](install_dependencies.sh) which could fix this
issue.  PRs are very welcome.


Former-commit-id: 5eefcece90ec8f1735db20fb56c40a8a3bac0db5